Replace old by new stream
