+++ ui/src/app/(dashboard)/stream/@slideOver/(.)trigger/[id]/page.tsx
