+++ ui/src/app/(dashboard)/stream/@slideOver/StreamDetails.tsx
