+++ ui/src/app/(dashboard)/stream/@slideOver/default.tsx
