+++ ui/src/app/(dashboard)/stream/OutputList.tsx
