+++ ui/src/app/(dashboard)/stream/SourceBadge.tsx
