+++ ui/src/app/(dashboard)/stream/layout.tsx
