+++ ui/src/app/(dashboard)/stream/trigger/[id]/page.tsx
